requirements.txt
certifi==2020.6.20
chardet==3.0.4
grpcio==1.30.0
grpcio-tools==1.30.0
idna==2.10
protobuf==3.12.2
six==1.15.0
urllib3==1.25.9
